requirements.txt
mysql-connector-python
pydantic
flask
requests
python-dotenv
flask-mysqldb
fastapi
uvicorn
boto3